New flow for download procedure with storing failures on fs

+++ packages/ckeditor5-dev-env/lib/translations/download.js
@@ -16,8 +16,10 @@ const logger = loggerFactory();
 const { cleanPoFileContent, createDictionaryFromPoFileContent } = translations;
 
 /**
- * Downloads translations from the Transifex for each localizable package.
- * It creates PO files out of the translations and replaces old translations with the downloaded ones.
+ * Downloads translations from the Transifex for each localizable package. It creates `*.po` files out of the translations and replaces old
+ * translations with the downloaded ones. If not all translations have been downloaded successfully, the `.transifex-failed-downloads.json`
+ * file is created, containing information about the packages and languages for which the translations could not be downloaded. This file is
+ * then used next time this script is run: it will try to download translations only for packages and languages that failed previously.
  *
  * @param {Object} config
  * @param {String} config.organizationName Name of the organization to which the project belongs.
@@ -44,19 +46,32 @@ module.exports = async function downloadTranslations( config ) {
 		localizablePackageNames
 	);
 
-	logger.info( 'Downloading translations...' );
+	const failedDownloads = [];
+	const { resourcesToProcess, hasAnyFailedDownload } = getResourcesToProcess( { cwd: config.cwd, resources, languages } );
 
-	for ( const resource of resources ) {
+	logger.info( hasAnyFailedDownload ?
+		'Downloading only translations that failed previously...' :
+		'Downloading all translations...'
+	);
+
+	for ( const { resource, languages } of resourcesToProcess ) {
 		const packageName = transifexService.getResourceName( resource );
-		const spinner = tools.createSpinner( `Processing "${ packageName }"...`, { indentLevel: 1 } );
 		const packagePath = config.packages.get( packageName );
 		const pathToTranslations = path.join( config.cwd, packagePath, 'lang', 'translations' );
+		const spinner = tools.createSpinner( `Processing "${ packageName }"...`, { indentLevel: 1 } );
 
 		spinner.start();
-		const translations = await transifexService.getTranslations( resource, languages );
 
-		// Remove all old translations before saving new ones.
-		fs.removeSync( pathToTranslations );
+		// Remove all old translations before saving new ones, but only if previously the download procedure has been finished without any
+		// failures. Otherwise, the current download procedure only tries to fetch the previously failed translations, so no existing files
+		// are removed beforehand.
+		if ( !hasAnyFailedDownload ) {
+			fs.removeSync( pathToTranslations );
+		}
+
+		const { translations, failedDownloads: failedDownloadsForPackage } = await transifexService.getTranslations( resource, languages );
+
+		failedDownloads.push( ...failedDownloadsForPackage );
 
 		const savedFiles = saveNewTranslations( {
 			pathToTranslations,
@@ -65,10 +80,21 @@ module.exports = async function downloadTranslations( config ) {
 		} );
 
 		spinner.finish();
-		logger.info( ' '.repeat( 3 ) + `✅ Saved ${ savedFiles } "*.po" files.` );
+
+		const statusMessage = failedDownloadsForPackage.length ?
+			`❌ Saved ${ savedFiles } "*.po" files. Some requests failed.` :
+			`✅ Saved all ${ savedFiles } "*.po" files.`;
+
+		logger.info( ' '.repeat( 3 ) + statusMessage );
 	}
 
-	logger.info( 'Saved all translations.' );
+	updateFailedDownloads( { cwd: config.cwd, failedDownloads } );
+
+	logger.info( failedDownloads.length ?
+		'Not all translations are saved. See `.transifex-failed-downloads.json` for details.\n' +
+		'Run the script again to fetch again only those translations that have failed to download.' :
+		'Saved all translations.'
+	);
 };
 
 /**
@@ -105,6 +131,85 @@ function saveNewTranslations( { pathToTranslations, translations, simplifyLicens
 	}
 
 	return savedFiles;
+}
+
+/**
+ * Based on whether previous download procedure has been finished without any failures, returns a collection of package names and language
+ * codes, for which translations will be downloaded:
+ *
+ * (1) If previous download procedure ended successfully, all translations for all resources will be downloaded.
+ * (2) Otherwise, only packages and their failed translation downloads defined in `.transifex-failed-downloads.json` are taken into account.
+ *
+ * @param {Object} config
+ * @param {String} config.cwd Current work directory.
+ * @param {Array.<Object>} config.resources All found resource instances for which translations could be downloaded.
+ * @param {Array.<Object>} config.languages All found language instances in the project.
+ * @returns {Object} result
+ * @returns {Boolean} result.hasAnyFailedDownload Indicates whether previous download procedure did not fetch all translations.
+ * @returns {Array.<Object>} result.resourcesToProcess Resource instances and their associated language instances to use during downloading
+ * the translations.
+ */
+function getResourcesToProcess( { cwd, resources, languages } ) {
+	const pathToFailedDownloads = path.join( cwd, '.transifex-failed-downloads.json' );
+	const hasAnyFailedDownload = fs.existsSync( pathToFailedDownloads );
+
+	if ( !hasAnyFailedDownload ) {
+		return {
+			hasAnyFailedDownload,
+			resourcesToProcess: resources.map( resource => ( { resource, languages } ) )
+		};
+	}
+
+	const resourcesMap = new Map( [
+		...resources.map( resource => [ transifexService.getResourceName( resource ), resource ] )
+	] );
+
+	const languagesMap = new Map( [
+		...languages.map( language => [ transifexService.getLanguageCode( language ), language ] )
+	] );
+
+	return {
+		hasAnyFailedDownload,
+		resourcesToProcess: fs.readJsonSync( pathToFailedDownloads )
+			.map( item => ( {
+				resource: resourcesMap.get( item.resourceName ),
+				languages: item.languages
+					.filter( language => languagesMap.has( language.code ) )
+					.map( language => languagesMap.get( language.code ) )
+			} ) )
+			.filter( item => item.resource && item.languages.length )
+	};
+}
+
+/**
+ * Saves all the failed downloads to `.transifex-failed-downloads.json` file. If there are no failures, the file is removed.
+ *
+ * @param {Object} config
+ * @param {String} config.cwd Current work directory.
+ * @param {Array.<Object>} config.failedDownloads Collection of all the failed downloads.
+ */
+function updateFailedDownloads( { cwd, failedDownloads } ) {
+	const pathToFailedDownloads = path.join( cwd, '.transifex-failed-downloads.json' );
+
+	if ( failedDownloads.length ) {
+		const groupedFailedDownloads = failedDownloads.reduce( ( result, failedDownload ) => {
+			const failedPackage = result.get( failedDownload.resourceName ) || {
+				resourceName: failedDownload.resourceName,
+				languages: []
+			};
+
+			failedPackage.languages.push( {
+				code: failedDownload.languageCode,
+				errorMessage: failedDownload.errorMessage
+			} );
+
+			return result.set( failedDownload.resourceName, failedPackage );
+		}, new Map() );
+
+		fs.writeJsonSync( pathToFailedDownloads, [ ...groupedFailedDownloads.values() ], { spaces: 2 } );
+	} else {
+		fs.removeSync( pathToFailedDownloads );
+	}
 }
 
 /**
